Fix deepedit transforms etc..
